(#678)    code		update reward amount type

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/remote/UserInfoClient.java
@@ -291,7 +291,7 @@ public interface UserInfoClient {
      */
     @FormUrlEncoded
     @POST(APP_PATH_REWARD_USER)
-    Observable<Object> rewardUser(@Path("user_id") long user_id, @Field("amount") float amount);
+    Observable<Object> rewardUser(@Path("user_id") long user_id, @Field("amount") Integer amount);
 
 
 }
